Removed UDE
Trying to emulate the device in the target is a horrible idea. The only good thing that came of that was being able to read debug messages from the driver... I lost all my notes on that when a hard drive crashed anyway. If I really want to nab those debug messages, I'll just patch the target driver and pipe them out through qemu somehow.